#101: Split AtlasMockAPI endoints to one endpoint per file
