+++ AtlasMockAPI/AtlasMockAPI/JSON/GET*|config.json
@@ -6,6 +6,6 @@
         "url": "http://localhost:9080"
     },
     "oauth2-provider": {
-        "url": "http://localhost:9080/oauth2/authorize?realm=/customers&response_type=token&client_id=CLIENT_ID&redirect_uri=http://com.atlas.app/redirect"
+        "url": "https://planb-provider-test.greendale-questionmark.zalan.do/oauth2/authorize?realm=/customers&response_type=token&client_id=atlas_Y2M1MzA&redirect_uri=http://de.zalando.atlas.AtlasCheckoutDemo/redirect"
     }
 }
